refactor: extract useListAction composable + VariantCardGrid component (#431)
Phase 1 continued — two more deduplication extractions:

- composables/useListAction.ts: shared action execution pattern
  (guard → confirm → launch check → telemetry → progress/run → result)
  extracted from DashboardView and InstallationList (~70 lines each)
- components/VariantCardGrid.vue: variant/device selection card grid
  extracted from QuickInstallModal, NewInstallModal, LoadSnapshotModal,
  and ModalDialog (4 identical template copies → 1 component)

Amp-Thread-ID: https://ampcode.com/threads/T-019db7b6-3400-7293-a487-351dcca1977b
