+++ src/renderer/src/views/DashboardView.vue
@@ -4,12 +4,9 @@ import { useI18n } from 'vue-i18n'
 import { useInstallationStore } from '../stores/installationStore'
 import { useSessionStore } from '../stores/sessionStore'
 import { useModal } from '../composables/useModal'
-import { useActionGuard } from '../composables/useActionGuard'
-import { useLocalInstanceGuard } from '../composables/useLocalInstanceGuard'
+import { useListAction } from '../composables/useListAction'
 import { useLauncherPrefs } from '../composables/useLauncherPrefs'
 import { useInstallContextMenu } from '../composables/useInstallContextMenu'
-import { emitTelemetryAction, toErrorBucket } from '../lib/telemetry'
-import { REQUIRES_STOPPED } from '../types/ipc'
 import { Download, Star, Clock, Cloud, Pin } from 'lucide-vue-next'
 import DashboardCard from '../components/DashboardCard.vue'
 import MigrationBanner from '../components/MigrationBanner.vue'
@@ -24,8 +21,6 @@ const { t } = useI18n()
 const installationStore = useInstallationStore()
 const sessionStore = useSessionStore()
 const modal = useModal()
-const actionGuard = useActionGuard()
-const localInstanceGuard = useLocalInstanceGuard()
 const prefs = useLauncherPrefs()
 
 const emit = defineEmits<{
@@ -233,77 +228,14 @@ function timeAgo(timestamp: number): string {
 }
 
 // --- Launch handling ---
+const { executeAction } = useListAction('dashboard', {
+  showProgress: (opts) => emit('show-progress', opts),
+})
+
 async function handleLaunch(inst: Installation, actions: ListAction[]): Promise<void> {
   const action = actions.find((a) => a.style === 'primary') ?? actions[0] ?? null
   if (!inst || !action) return
-  const telemetryContext = {
-    source_category: inst.sourceCategory || 'unknown',
-    ui_surface: 'dashboard',
-  }
-
-  if (action.enabled === false && action.disabledMessage) {
-    await modal.alert({ title: action.label, message: action.disabledMessage })
-    return
-  }
-
-  // Pre-flight: check if the installation is busy or running
-  if (REQUIRES_STOPPED.has(action.id)) {
-    if (!await actionGuard.checkBeforeAction(inst.id, action.label)) return
-  }
-
-  if (action.confirm) {
-    const confirmed = await modal.confirm({
-      title: action.confirm.title || 'Confirm',
-      message: action.confirm.message || 'Are you sure?',
-      confirmLabel: action.label,
-      confirmStyle: action.style || 'danger',
-    })
-    if (!confirmed) {
-      emitTelemetryAction('launcher.action.result', { action_id: action.id, result: 'cancelled', ...telemetryContext })
-      return
-    }
-  }
-
-  if (action.id === 'launch') {
-    const canLaunch = await localInstanceGuard.checkBeforeLaunch(inst.id)
-    if (!canLaunch) {
-      emitTelemetryAction('launcher.action.result', { action_id: action.id, result: 'cancelled', ...telemetryContext })
-      return
-    }
-  }
-
-  sessionStore.clearErrorInstance(inst.id)
-  emitTelemetryAction('launcher.action.invoked', { action_id: action.id, ...telemetryContext })
-  if (action.showProgress) {
-    emit('show-progress', {
-      installationId: inst.id,
-      title: `${action.progressTitle || action.label} — ${inst.name}`,
-      apiCall: () => window.api.runAction(inst.id, action.id),
-      cancellable: !!action.cancellable,
-    })
-    return
-  }
-
-  try {
-    const result = await window.api.runAction(inst.id, action.id)
-    if (result.running) {
-      await actionGuard.checkBeforeAction(inst.id, action.label)
-      return
-    }
-    const resultValue = result.cancelled ? 'cancelled' : (result.ok === false ? 'failed' : 'ok')
-    emitTelemetryAction('launcher.action.result', { action_id: action.id, result: resultValue, ...telemetryContext })
-    if (result.message) {
-      await modal.alert({ title: action.label, message: result.message })
-    }
-  } catch (error: unknown) {
-    emitTelemetryAction('launcher.action.result', {
-      action_id: action.id,
-      result: 'failed',
-      error_bucket: toErrorBucket(error),
-      ...telemetryContext,
-    })
-    throw error
-  }
+  await executeAction(inst, action)
 }
 
 // --- Change primary ---
