+++ src/renderer/src/views/InstallationList.vue
@@ -4,13 +4,10 @@ import { useI18n } from 'vue-i18n'
 import { useSessionStore } from '../stores/sessionStore'
 import { useInstallationStore } from '../stores/installationStore'
 import { useModal } from '../composables/useModal'
-import { useActionGuard } from '../composables/useActionGuard'
-import { useLocalInstanceGuard } from '../composables/useLocalInstanceGuard'
+import { useListAction } from '../composables/useListAction'
 import { useInstallContextMenu } from '../composables/useInstallContextMenu'
 import { useProgressStore } from '../stores/progressStore'
 import { DraggableList } from '../lib/draggableList'
-import { emitTelemetryAction, toErrorBucket } from '../lib/telemetry'
-import { REQUIRES_STOPPED } from '../types/ipc'
 import InstanceCard from '../components/InstanceCard.vue'
 import ContextMenu from '../components/ContextMenu.vue'
 import type { Installation, ListAction } from '../types/ipc'
@@ -20,8 +17,6 @@ const sessionStore = useSessionStore()
 const installationStore = useInstallationStore()
 const progressStore = useProgressStore()
 const modal = useModal()
-const actionGuard = useActionGuard()
-const localInstanceGuard = useLocalInstanceGuard()
 
 const filter = ref('all')
 const listActions = ref(new Map<string, ListAction[]>())
@@ -111,75 +106,23 @@ function getLaunchMeta(inst: Installation): string {
   return ''
 }
 
-async function handleListAction(inst: Installation, action: ListAction): Promise<void> {
-  const telemetryContext = {
-    source_category: inst.sourceCategory || 'unknown',
-    ui_surface: 'list',
-  }
-  if (action.enabled === false && action.disabledMessage) {
-    await modal.alert({ title: action.label, message: action.disabledMessage })
-    return
-  }
-  if (inst.seen === false) {
-    inst.seen = true
-    window.api.updateInstallation(inst.id, { seen: true })
-  }
-  // Pre-flight: check if the installation is busy or running
-  if (REQUIRES_STOPPED.has(action.id)) {
-    if (!await actionGuard.checkBeforeAction(inst.id, action.label)) return
-  }
-  if (action.confirm) {
-    const confirmed = await modal.confirm({
-      title: action.confirm.title || 'Confirm',
-      message: action.confirm.message || 'Are you sure?',
-      confirmLabel: action.label,
-      confirmStyle: action.style || 'danger',
-    })
-    if (!confirmed) {
-      emitTelemetryAction('launcher.action.result', { action_id: action.id, result: 'cancelled', ...telemetryContext })
-      return
-    }
-  }
-  if (action.id === 'launch') {
-    const canLaunch = await localInstanceGuard.checkBeforeLaunch(inst.id)
-    if (!canLaunch) {
-      emitTelemetryAction('launcher.action.result', { action_id: action.id, result: 'cancelled', ...telemetryContext })
-      return
-    }
-  }
-  sessionStore.clearErrorInstance(inst.id)
-  emitTelemetryAction('launcher.action.invoked', { action_id: action.id, ...telemetryContext })
-  if (action.showProgress) {
-    emit('show-progress', {
-      installationId: inst.id,
-      title: `${action.progressTitle || action.label} — ${inst.name}`,
-      apiCall: () => window.api.runAction(inst.id, action.id),
-      cancellable: !!action.cancellable,
-    })
-    return
-  }
-  try {
-    const result = await window.api.runAction(inst.id, action.id)
-    if (result.running) {
-      await actionGuard.checkBeforeAction(inst.id, action.label)
-      return
-    }
-    const resultValue = result.cancelled ? 'cancelled' : (result.ok === false ? 'failed' : 'ok')
-    emitTelemetryAction('launcher.action.result', { action_id: action.id, result: resultValue, ...telemetryContext })
+const { executeAction } = useListAction('list', {
+  showProgress: (opts) => emit('show-progress', opts),
+  onNavigate: async (result, action) => {
     if (result.navigate === 'list') {
       await refresh()
     } else if (result.message) {
       await modal.alert({ title: action.label, message: result.message })
     }
-  } catch (error: unknown) {
-    emitTelemetryAction('launcher.action.result', {
-      action_id: action.id,
-      result: 'failed',
-      error_bucket: toErrorBucket(error),
-      ...telemetryContext,
-    })
-    throw error
+  },
+})
+
+async function handleListAction(inst: Installation, action: ListAction): Promise<void> {
+  if (inst.seen === false) {
+    inst.seen = true
+    window.api.updateInstallation(inst.id, { seen: true })
   }
+  await executeAction(inst, action)
 }
 
 // --- Drag-to-reorder ---
