+++ src/renderer/src/views/QuickInstallModal.vue
@@ -5,7 +5,8 @@ import { useModal } from '../composables/useModal'
 import { useModalOverlay } from '../composables/useModalOverlay'
 import type { Source, FieldOption } from '../types/ipc'
 import { emitTelemetryAction, toVariantBucket } from '../lib/telemetry'
-import { stripVariantPrefix, getVariantImage, sortedCardOptions } from '../lib/variants'
+import { stripVariantPrefix, sortedCardOptions } from '../lib/variants'
+import VariantCardGrid from '../components/VariantCardGrid.vue'
 import { formatBytes } from '../lib/formatting'
 import { toPathGuardrail, trackGuardrailBlocked, trackDiskWarningResponse, createDiskSpaceChecker } from '../lib/installHelpers'
 
@@ -358,34 +359,11 @@ defineExpose({ open })
 
             <div class="field">
               <label>{{ $t('quickInstall.selectVariant') }}</label>
-              <div class="variant-cards">
-                <div
-                  v-for="opt in sortedCardOptions(variantOptions)"
-                  :key="opt.value"
-                  :class="['variant-card', {
-                    selected: selectedVariant?.value === opt.value,
-                    recommended: opt.recommended
-                  }]"
-                  @click="selectVariant(opt)"
-                >
-                  <div class="variant-card-icon">
-                    <img
-                      v-if="getVariantImage(opt)"
-                      :src="getVariantImage(opt)!"
-                      :alt="opt.label"
-                      draggable="false"
-                    />
-                    <span v-else class="variant-card-icon-text">{{ opt.label }}</span>
-                  </div>
-                  <div class="variant-card-label">{{ opt.label }}</div>
-                  <div v-if="opt.recommended" class="variant-card-badge">
-                    {{ $t('newInstall.recommended') }}
-                  </div>
-                  <div v-if="opt.description" class="variant-card-desc">
-                    {{ opt.description }}
-                  </div>
-                </div>
-              </div>
+              <VariantCardGrid
+                :options="sortedCardOptions(variantOptions)"
+                :selected-value="selectedVariant?.value"
+                @select="selectVariant"
+              />
             </div>
 
             <div class="field">
